feat(dev): Add support for masking arrays
The `maskObject` function is used to transform controller state in a way
that prevents private data from reaching Sentry. The way it generally
works is that you pass a second object that has the same shape as the
object you want to transform but contains `true` or `false` in place of
the corresponding values.

If you want to transform all properties of an object in the same way,
you can create a special object in the mask with an `AllProperties`
property, and then you specify a submask for the value. Unfortunately it
is impossible to transform all values of an array using the same
technique. This commit makes that possible.

Adding support for this feature was much easier when the existing file
was converted to TypeScript, so that has been done as well. In addition,
as the complexity increased, unit tests have been added to ensure that
the existing logic still works.