test: Remove feature flags from onboarding fixture
Feature flags were accidentally added to the onboarding fixture
recently. This wasn't a problem for E2E tests because we refresh
feature flags as part of initialization (which itself was a bug), but
when we tried to fix this in #37552 it revealed this fixture problem.

The onboarding fixture has been updated to not include any feature
flags. Individual E2E tests can set the features they need using a
manifest override or a feature flag API mock.