fix: Fix broken error capture
The hook `useDappSwapComparisonInfo` was using the `@sentry/browser`
import `captureException` to capture exceptions in one failure
condition. This fails 100% of the time in production builds because
the Sentry library is not initialized in this execution context
(it's initialized in the global environment, but not in the root
Compartment that our application code runs in).

This also had the effect of adding `@sentry/browser` to the UI bundle,
which makes the bundle larger and slower to parse for no benefit.

The error handling has been updated to use the shared
`captureException` utility function that correctly calls the Sentry
instance that has been initialized. This ensures the error is captured
correctly, and reduces the bundle size.