chore: improve getMetaMaskAccounts performance
This selector depends on so much state that on re-renders
we hardly hit the memoization, and instead it ends up
being recomputted multiple times. That comes with a performance
impact, as the isAddressCompatible computations are
quite expensive and we are running them multiple times
for each account (we are trying out every possible non evm
chain until we have a match. This times the total number of
accounts (solana + btc + tron).
This commit suggests another approach, where we can identify
the account by cross referencing its type with the network
providers chainId.
On a local build the performance of the application greatly
improved (for example on of the long running tasks on the
popup initialisation went from 1.8s to take approx 180ms)