chore: Remove `extension-devs` CODEOWNER for LavaMoat policies
We have a new team (`policy-reviwers`) specifically for those
responsible for reviewing LavaMoat policies. The `extension-devs` team
was initially left in place to help transition to the new policy
reviewers team, but it has been around long enough now that everyone
who should be a reviewer is on the team (and if not, we'll quickly
find out).