chore: Increase UI<=>background RPC timeout
The timeout for the `getState` call from the UI to the background has
been increased from 10s to 16s. This increase may be enough to allow
state to load on some slower machines where the wallet state is
especially large, and at the very least if we don't see a change in
error rates, we'll still learn something from this change.

This change is a compainion to #38322, which increased a different
related timeout for the same reason.