chore: Reorganize ESLint config
The ESLint configuration file was intended to be sorted into particular
groups, but over time we've deviated from this. The configuration file
has been updated to restore the grouping we used before, roughly.

Groups have been considated where possible, and redundant rules/config
has been removed.

This should result in no additional lint errors, and no changes in how
many warnings there are. The actual changes to which rules are applied
to any given file should be minimal and inconsequential.