chore: Remove duplicate Sentry snapshot keys
It seems that these snapshots have been updated manually with duplicate
keys, which were ignored when the snapshots were read by the test.

I was able to eliminate the duplicates by manually making a change to
each snapshot so they are flagged as outdated, then regenerating them
by running the test with `UPDATE_SNAPSHOTS=true`.