test: Prevent double `FixtureExtensionStore` initialization
The `FixtureExtensionStore` gets initialized both in the background and
in each UI instance. Each time it reads the fixture state from disk and
writes it to `storage.local`.

The store has been updated to optionally skip initialization. It no
longer initializes by default, you need to opt-in to that step. It will
now only initialize in the background process.